Added files via upload
This version fix the following problem: 

After link 121, the code fails to find another link to further reduce the objective function. Therefore, the code selects link 121 as the service candidate and tries to remove it from the non-serviced required links. And that cause an error. The error is aroused because the service_finished_time is not updated properly.